Don't crash when type is :selector_popover on iPhone
When using type: :selector_popover on iPhone the app crashes. This PR checks the type and device idiom and falls back to :selector_push if the device is an iPhone and the user is trying to use :selector_popover. It also outputs an error message in development mode if they're trying to do this.

Conflicts:
	lib/ProMotion/XLForm/xl_form_cell_builder.rb